Ballot FORUM-XX: Allow relative effective dates in ballots
Allow the Chair or Vice Chair of the Forum or a Working Group to update relative effective dates found within a ballot into concrete dates when publishing a Final Guideline or Final Maintenance Guideline.

This allows ballots to contain statements like "Effective [6 months after publication], CAs MUST...", which the Chair would then convert to "Effective Jan 15, 2026, CAs MUST". This allows the ballot writing and discussion process to focus on the length of the necessary transition/implementation period, rather than on the specific effective date, preventing ballots from needing to be re-proposed simply because the discussion period has taken longer than expected and now the effective date is too close.